💬 修复 mybatis plus 在 jdk 17 不兼容问题

+++ backend/src/main/java/com/pearadmin/pro/modules/sys/service/impl/SysUserServiceImpl.java
@@ -110,7 +110,7 @@ public class SysUserServiceImpl extends ServiceImpl<SysUserRepository, SysUser> 
 
     @Override
     public Boolean exist(String username) {
-        int count = this.lambdaQuery().eq(SysUser::getUsername, username).count();
+        long count = this.lambdaQuery().eq(SysUser::getUsername, username).count();
         if(count > 0) {
             return true;
         } else {
